wip: moving stats for span impl to sql server
