+++ pkg/sql/logictest/testdata/logic_test/pg_builtins
@@ -129,7 +129,7 @@ SELECT pg_table_is_visible(1010101010)
 NULL
 
 statement ok
-CREATE TABLE size_tester (a PRIMARY KEY, b) AS SELECT g,g+1 FROM generate_series(1,100) g(g)
+CREATE TABLE size_tester (a PRIMARY KEY, b string) AS SELECT g, 'preston' FROM generate_series(1,100) g(g)
 
 # In tests, this returns NULL because ranges don't appear properly in
 # crdb_internal.ranges in the logic test context.
